kafka/cons-requirements.txt
confluent-kafka
pandas
pyspark
delta-spark
numpy